celestia-cloudrun-pipeline/backend/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
httpx==0.27.2
python-dotenv==1.0.1
loguru==0.7.2
google-cloud-tasks==2.16.6
